add Go linters using golangci-lint
This was based off, initially, the letsencrypt/boulder golangci-lint
config, but made simpler by allowing more checks and ignoring fewer
APIs. This has led to a number of other PRs to correct lint errors.
Those are catalogued in the comments of #294.

Multiple .golangci.yml files have been added because the golangci-lint
Action doesn't seem to have a documented argument to change the config
file path. So, we needed to add a config file for each Go subproject.

Similarly, there's an entry for each directory in the Github Action CI
config because golangci-lint needs a `go.mod` in the same directory to
work correctly and cannot operate on multiple Go modules at once.

One controversial change from the letsencrypt/boulder config is letting
gosec error on what it believes may be hard-coded secrets (G101 in gosec
lingo).

Fixes #294